backend/pyproject.toml
[tool.poetry]
name = "skylock"
version = "0.1.0"
description = ""
authors = ["akempins <01178535@pw.edu.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.3"
pylint = "^3.3.1"
fastapi = "^0.115.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.0"
bcrypt = "^4.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
